pyproject.toml
[project]
name = "bender"
version = "0.1.1"
description = "Data bending toolkit"
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "click>=8.1.8",
    "librosa>=0.10.2.post1",
    "numba>=0.61.0",
    "numexpr>=2.11.0",
    "numpy>=2.1.3",
    "pillow>=11.1.0",
    "soundfile>=0.13.1",
    "watchdog>=6.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]

[project.scripts]
bender = "bender:__main__.main"

[tool.mypy]
plugins = "numpy.typing.mypy_plugin"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning"
]
